+++ CHANGES.md
@@ -1,6 +1,11 @@
  
 # [Quark Engine Java ](https://quark.greenscreens.ltd/).
 
+## v7.0.0. (Preview)
+ - Add Java 21 support
+ - Switch from ThreadLocal to JEP 429 Scoped Variables
+ - Add support for JEP 436 Virtual Threads
+
 ## v6.0.0. (2023.Q4)
  - Switch to Jakarta EE
  - Complete Quark code refactoring